@@ -7,38 +7,36 @@
     "license": "MIT",
     "authors": [
         {
-            "name": "Bohan Yang",
-            "email": "youthdna@live.com"
+            "name": "Brent, Bohan Yang",
+            "email": "brent@manyou.io"
         }
     ],
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
         "chubbyphp/chubbyphp-workerman-request-handler": "^2.0",
         "nyholm/psr7": "^1.5",
         "symfony/psr-http-message-bridge": "^2.1",
-        "symfony/runtime": "^5.4 || ^6.0",
+        "psr/http-message": "^1.0",
+        "symfony/runtime": "^6.1",
         "workerman/workerman": "^4.0",
-        "symfony/http-kernel": "^5.4 || ^6.0"
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "config": {
-        "allow-plugins": {
-            "symfony/runtime": true,
-            "dealerdirect/phpcodesniffer-composer-installer": true
-        }
+        "symfony/http-kernel": "^6.1",
+        "symfony/http-foundation": "^6.1",
+        "psr/http-factory": "^1.0",
+        "psr/http-server-handler": "^1.0",
+        "psr/log": "^2.0 || ^3.0"
     },
     "require-dev": {
-        "doctrine/coding-standard": "9.0.x-dev",
-        "slevomat/coding-standard": "^8.1",
-        "squizlabs/php_codesniffer": "^3.7",
-        "webimpress/coding-standard": "^1.2",
-        "blackfire/php-sdk": "^1.31",
-        "monolog/monolog": "^3.1"
+        "blackfire/php-sdk": "^1.31"
+    },
+    "minimum-stability": "dev",
+    "extra": {
+        "branch-alias": {
+            "dev-main": "0.1-dev"
+        }
     },
     "autoload": {
         "psr-4": {
-            "Manyou\\WorkermanSymfonyRuntime\\": "src/"
+            "Manyou\\WorkermanSymfonyRuntime\\": ""
         }
     }
 }
